comment removement, auth validation before actions added

+++ src/service/CommentLikeService.js
@@ -3,7 +3,11 @@ import axios from 'axios';
 function addCommentLike(commentId) {
     const accessToken = localStorage.getItem('accessToken');
     const refreshToken = localStorage.getItem('refreshToken');
-    const userId = localStorage.getItem('userId');
+    const userId = localStorage.getItem('userId') || "";
+
+    if (userId === "") {
+        return null;
+    }
 
     return axios.post("http://localhost:8080/api/comment-like/create", { commentId, userId }, {
         headers: {
@@ -18,13 +22,18 @@ function addCommentLike(commentId) {
     })
     .catch(error => {
         console.log(error);
+        return null;
     });
 }
 
 function removeCommentLike(commentId) {
     const accessToken = localStorage.getItem('accessToken');
     const refreshToken = localStorage.getItem('refreshToken');
-    const userId = localStorage.getItem('userId');
+    const userId = localStorage.getItem('userId') || "";
+
+    if (userId === "") {
+        return null;
+    }
 
     return axios.delete("http://localhost:8080/api/comment-like/delete", { params: { commentId, userId }, headers: {
         'Accept': 'application/json',
@@ -37,6 +46,7 @@ function removeCommentLike(commentId) {
     })
     .catch(error => {
         console.log(error);
+        return null;
     });
 }
